Refactor SoftMig checks and improve logging in CUDA memory functions
- Simplified the SoftMig check in `cuDeviceTotalMem_v2` by removing redundant shared region validation, as it is already handled by `is_softmig_enabled()`.
- Enhanced logging in `cuMemMap` to ensure the offset is logged correctly as an unsigned long long, improving clarity in memory mapping diagnostics.
- Added forward declarations for external functions in `multiprocess_memory_limit.c` to maintain code organization and readability.

+++ src/cuda/memory.c
@@ -701,7 +701,7 @@ CUresult cuMemMap( CUdeviceptr ptr, size_t size, size_t offset, CUmemGenericAllo
     if (size > 1024*1024) {
         LOG_INFO("cuMemMap: size=%.2f GB ptr=%llx", size / (1024.0 * 1024.0 * 1024.0), ptr);
     } else {
-        LOG_DIAG("cuMemMap: size=%lu ptr=%llx offset=%llx", size, ptr, offset);
+        LOG_DIAG("cuMemMap: size=%lu ptr=%llx offset=%llx", size, ptr, (unsigned long long)offset);
     }
     CUresult res = CUDA_OVERRIDE_CALL(cuda_library_entry,cuMemMap,ptr,size,offset,handle,flags);
     return res;
